+++ estate-bill/src/main/java/com/estate/sdzy/service/impl/SRoleMenuServiceImpl.java
@@ -1,14 +1,18 @@
 package com.estate.sdzy.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.estate.exception.BillException;
 import com.estate.sdzy.entity.SRoleMenu;
 import com.estate.sdzy.mapper.SRoleMenuMapper;
 import com.estate.sdzy.service.SRoleMenuService;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import com.estate.util.BillExceptionEnum;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
 
 /**
  * <p>
@@ -32,22 +36,25 @@ public class SRoleMenuServiceImpl extends ServiceImpl<SRoleMenuMapper, SRoleMenu
         String[] menuIdArr = menuIds.split(",");
 
         QueryWrapper<SRoleMenu> query = new QueryWrapper<>();
-        query.eq("role_id",roleId);
-        try{
-            // 执行修改之前，先将所有的该角色的菜单关系删除，然后在从新添加
-            roleMenuMapper.delete(query);
-
-            for (String menuId : menuIdArr) {
-                SRoleMenu roleMenu = new SRoleMenu();
-                roleMenu.setMenuId(Long.valueOf(menuId));
-                roleMenu.setRoleId(roleId);
-//            roleMenu.setCompId();
-                roleMenuMapper.insert(roleMenu);
+        query.eq("role_id", roleId);
+        // 执行修改之前，先将所有的该角色的菜单关系删除，然后在从新添加
+        List<SRoleMenu> sRoleMenus = roleMenuMapper.selectList(query);
+        if(!sRoleMenus.isEmpty()){
+            int delete = roleMenuMapper.delete(query);
+            if(!(delete > 0)){
+                throw new BillException(BillExceptionEnum.SET_ROLE_MENU_ERROR);
             }
-        }catch (Exception e){
-            e.printStackTrace();
-            log.error("删除菜单出现错误，角色id={}",roleId);
-            return  false;
+        }
+
+        for (String menuId : menuIdArr) {
+            SRoleMenu roleMenu = new SRoleMenu();
+            roleMenu.setMenuId(Long.valueOf(menuId));
+            roleMenu.setRoleId(roleId);
+//            roleMenu.setCompId();
+            int insert = roleMenuMapper.insert(roleMenu);
+            if(!(insert>0)){
+                throw new BillException(BillExceptionEnum.SET_ROLE_MENU_ERROR);
+            }
         }
 
         return true;
