+++ estate-bill/src/main/java/com/estate/sdzy/service/impl/SUserServiceImpl.java
@@ -71,6 +71,12 @@ public class SUserServiceImpl extends ServiceImpl<SUserMapper, SUser> implements
         user.setState(state);
 
         int insert = userMapper.insert(user);
+        if (insert > 0) {
+            log.info("自动添加公司管理员角色，登录用户名={}admin,密码=123456", userName);
+        } else {
+            log.error("自动添加管理员角色失败,请联系管理员添加账号");
+            throw new BillException(BillExceptionEnum.SYSTEM_INSERT_ERROR);
+        }
 
         return insert > 0;
     }
@@ -85,25 +91,28 @@ public class SUserServiceImpl extends ServiceImpl<SUserMapper, SUser> implements
         // 更新用户权限的时候，需要先删除之前存在的用户角色关系然后从新生成。
         QueryWrapper<SUserRole> queryWrapper = new QueryWrapper<>();
         queryWrapper.eq("user_id", userId);
-        int delete = userRoleMapper.delete(queryWrapper);
-
-        if (delete > 0) {
-            SUser user = getUserByToken(token);
-            String[] roleIdArr = roleIds.split(",");
-            for (String roleId : roleIdArr) {
-                SUserRole userRole = new SUserRole();
-                userRole.setUserId(userId);
-                userRole.setRoleId(Long.valueOf(roleId));
-                userRole.setCreatedBy(user.getId());
-                userRole.setCreatedName(user.getUserName());
-
-                int insert = userRoleMapper.insert(userRole);
-                if(!(insert > 0)){
-                    throw new BillException(BillExceptionEnum.SET_USER_ROLE_ERROR);
-                }
+        // 先查询用户的权限，如果没有权限，直接添加，如果有权限，需要先删除在添加新的权限
+        List<SUserRole> sUserRoles = userRoleMapper.selectList(queryWrapper);
+        if (!sUserRoles.isEmpty()) {
+            int delete = userRoleMapper.delete(queryWrapper);
+            if(!(delete > 0)){
+                throw new BillException(BillExceptionEnum.SET_USER_ROLE_ERROR);
             }
-        }else {
-            throw new BillException(BillExceptionEnum.SET_USER_ROLE_ERROR);
+        }
+
+        SUser user = getUserByToken(token);
+        String[] roleIdArr = roleIds.split(",");
+        for (String roleId : roleIdArr) {
+            SUserRole userRole = new SUserRole();
+            userRole.setUserId(userId);
+            userRole.setRoleId(Long.valueOf(roleId));
+            userRole.setCreatedBy(user.getId());
+            userRole.setCreatedName(user.getUserName());
+
+            int insert = userRoleMapper.insert(userRole);
+            if (!(insert > 0)) {
+                throw new BillException(BillExceptionEnum.SET_USER_ROLE_ERROR);
+            }
         }
 
         return true;
@@ -112,13 +121,13 @@ public class SUserServiceImpl extends ServiceImpl<SUserMapper, SUser> implements
     @Override
     public boolean save(SUser user, String token) {
         SUser users = getUserByToken(token);
-        if(null == user){
+        if (null == user) {
             throw new BillException(BillExceptionEnum.PARAMS_MISS_ERROR);
         }
         int insert = userMapper.insert(user);
         if (insert > 0) {
             log.info("用户添加成功，添加人={}", users.getUserName());
-        }else{
+        } else {
             throw new BillException(BillExceptionEnum.SYSTEM_INSERT_ERROR);
         }
         return insert > 0;
@@ -130,7 +139,7 @@ public class SUserServiceImpl extends ServiceImpl<SUserMapper, SUser> implements
         int i = userMapper.updateById(user);
         if (i > 0) {
             log.info("用户更新成功，修改人={}", users.getUserName());
-        }else {
+        } else {
             throw new BillException(BillExceptionEnum.SYSTEM_UPDATE_ERROR);
         }
         return i > 0;
@@ -142,7 +151,7 @@ public class SUserServiceImpl extends ServiceImpl<SUserMapper, SUser> implements
         int i = userMapper.deleteById(id);
         if (i > 0) {
             log.info("用户删除成功，删除人={}", users.getUserName());
-        }else{
+        } else {
             throw new BillException(BillExceptionEnum.SYSTEM_DELETE_ERROR);
         }
         return i > 0;
